REFACTO : added length constraints for username in register serializer
