+++ backend/modules/accounts/views/reset_password.py
@@ -4,7 +4,7 @@ from rest_framework import status
 from django.contrib.auth.tokens import default_token_generator
 from django.utils.http import urlsafe_base64_decode
 from django.contrib.auth import get_user_model
-from modules.accounts.serializers import ResetPasswordSerializer
+from modules.accounts.serializers.legacy_serializers import ResetPasswordSerializer
 from modules.accounts.auth_utils.cookie import attach_jwt_cookie
 
 User = get_user_model()
